% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_extra_info.R
\name{append_extra_info}
\alias{append_extra_info}
\title{append_extra_info}
\usage{
append_extra_info(fileOfRNA, fileOfATAC, fileOfpeakMap, fileOfgeneMap)
}
\arguments{
\item{fileOfRNA}{RNA sequencing data file}

\item{fileOfATAC}{ATAC sequencing data file}

\item{fileOfpeakMap}{peak map file}

\item{fileOfgeneMap}{gene map file}
}
\value{
Two dataframes, one for RNA data and one for ATAC data
}
\description{
Adding gene names and other information to RNA data,
 as well as integrating ATAC data
}
\examples{
\dontrun{
rna <- "TCGA-BRCA.htseq_fpkm-uq.tsv"
atac <- "brca_brca_peak_Log2Counts_dedup.brca_brca_peak_log2counts_dedup"
peakmap <- "brca_brca_peak.probeMap"
append_extra_info(rna,atac,peakmap)
}
}
