% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacmap.R
\docType{data}
\name{atacmap}
\alias{atacmap}
\title{Data frame for directly usable ATAC peak map}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 215920 rows and 6 columns.
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
atacmap
}
\description{
Data frame for directly usable ATAC peak map
}
\keyword{datasets}
