% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage}
\alias{linkage}
\title{Calculate correlation
An auxiliary function used to calculate
the correlation coefficient and p-value}
\usage{
linkage(dataOfRNA, dataOfATAC, geneset, rho = 0.3, p = 0.1)
}
\arguments{
\item{dataOfRNA}{data frame of RNA data}

\item{dataOfATAC}{data frame of ATAC data}

\item{geneset}{Gene vectors of interest}

\item{rho}{Correlation coefficient threshold}

\item{p}{p-value threshold}
}
\value{
A data frame. Include the following information

hgnc_symbol,chrom ,chromStart,chromEnd,idx_gene,idx_peak,rho,p_value
}
\description{
Calculate correlation
An auxiliary function used to calculate
the correlation coefficient and p-value
}
\examples{
\dontrun{linkage(RNA,ATAC,RNA$hgnc_symbol[1:10])}

}
