% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cor_test.R
\name{my_cor_test}
\alias{my_cor_test}
\title{Calculate correlation
An auxiliary function used to calculate
the correlation coefficient and p-value}
\usage{
my_cor_test(pair)
}
\arguments{
\item{pair}{Index Pairs of Genes and Peaks in Various Data Frames}
}
\value{
coefficient and p-value
}
\description{
Calculate correlation
An auxiliary function used to calculate
the correlation coefficient and p-value
}
