% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_anno.R
\name{peak_anno}
\alias{peak_anno}
\title{annotation peak
Annotate peak and visualize}
\usage{
peak_anno(atacmap)
}
\arguments{
\item{atacmap}{peak map file}
}
\value{
Annotate peak and visualize
}
\description{
annotation peak
Annotate peak and visualize
}
\examples{
\dontrun{peak_anno(atacmap)}

}
