% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R
\docType{data}
\name{testATAC}
\alias{testATAC}
\title{Data frame for directly usable ATAC data}
\format{
An object of class \code{data.frame} with 215920 rows and 75 columns.
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
testATAC
}
\description{
This dataset is an ATAC-seq study of GDC TCGA Breast Cancer (BRCA) cohort with
74 samples and 215,921 identifiers. The measurement unit is log2((count+5)PM)-qn for
all peak signal. ATAC-seq technology can help us better understand the open
chromatin regions in the genome, providing insights into the gene regulatory mechanisms
and potential treatments for breast cancer. This dataset provides comprehensive
ATAC-seq data that can assist researchers in analyzing gene expression profiles,
genomic structures, and other key features of breast cancer,
advancing breast cancer research and improving clinical treatment strategies.
}
\keyword{datasets}
