% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\docType{data}
\name{testRNA}
\alias{testRNA}
\title{Data frame for directly usable RNA data}
\format{
An object of class \code{data.frame} with 56339 rows and 77 columns.
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
testRNA
}
\description{
This dataset is a study of gene expression RNAseq for GDC TCGA Breast Cancer (BRCA) cohort.
It contains 1217 samples and 60,484 identifiers. The data was collected using Illumina platform
and HTSeq method to obtain RNAseq gene expression data for each sample.
The measurement unit is log2(fpkm-uq+1), which can be used to investigate the
gene expression profile of breast cancer patients, and to understand the biology,
disease mechanisms, and potential treatment options of breast cancer.
This dataset provides important reference material for breast cancer research,
and reliable data sources for relevant research in medical and life sciences.
}
\keyword{datasets}
